backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
mysql-connector-python==8.1.0
python-dotenv==1.0.1